feat: Add pagination to JobsTable
The JobsTable component now supports pagination. The page number and
size are passed as query parameters to the getJobs function. The
response from getJobs is expected to be an object containing the jobs
data, the current page, the page size, and the total number of jobs.
The Pagination component from antd is used to display the pagination
controls.